her/jax_her/requirements.txt
wandb
gymnasium-robotics
flax
optax
distrax
mpi4jax
ipython
tqdm
pandas
ml-collections
imageio-ffmpeg
gitpython